requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
Flask==1.1.2
gcloud==0.18.3
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jws==0.1.3
MarkupSafe==1.1.1
Naked==0.1.31
oauth2client==4.1.3
packaging==20.9
pluggy==0.13.1
protobuf==3.15.7
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase4==4.4.3
pytest==6.2.3
pytest-cov==2.11.1
python-jwt==2.0.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
shellescape==3.8.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
venmo_api==0.3.0
